requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cld2-cffi==0.1.4
click==7.1.2
decorator==4.4.2
distlib==0.3.1
Django==3.1
django-cors-headers==3.4.0
django-extensions==3.0.9
django-filter==2.3.0
django-guardian==2.3.0
django-multiupload==0.5.2
django-pandas==0.6.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
ed25519==1.5
filelock==3.0.12
Flask==1.1.2
gensim==3.8.3
gunicorn==20.0.4
idna==2.10
install==1.3.3
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
multi-rake==0.0.1
nltk==3.5
numpy==1.19.1
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-rake==1.4.5
pytz==2020.1
rake-nltk==1.0.4
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
smart-open==2.2.1
sqlparse==0.3.1
Token==1.8.6
tokenizer==2.1.0
toml==0.10.1
tqdm==4.49.0
traitlets==4.3.3
urllib3==1.25.10
virtualenv==20.0.30
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
